little fix in Triangle calculation and finished persons

+++ src/main/java/io/codelex/oop/shapes/Rectangle.java
@@ -4,7 +4,7 @@ public class Rectangle extends Shape {
     private double width;
     private double height;
 
-    public Rectangle(int numSides, int width, int height) {
+    public Rectangle(int numSides, double width, double height) {
         super(numSides);
         this.width = width;
         this.height = height;
